refactor(manage-list): split AddStudentDialog into modular components
- Improved overall code readability and maintainability
- Created reusable subcomponents for form fields, header, and footer
- Added dedicated form validation logic in utils and hooks
- Organized files under add-dialog folders in components, hooks, types, and constants
